refactor: Move RegisterTrackingUseCase to domain layer (#129)
